requirements-runtime.txt
typer==0.12.3
PyYAML==6.0.2
psycopg2-binary==2.9.9
